+++ my_solved/lib/features/splash/bloc/splash_state.dart
