+++ my_solved/lib/features/splash/screen/splash_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:my_solved/features/splash/splash_bloc.dart';
+import 'package:my_solved/features/splash/bloc/splash_bloc.dart';
 
 class SplashScreen extends StatelessWidget {
   const SplashScreen({super.key});
